+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - PR #3052: Speeding up MNMG KNN Cl&Re testing
 - PR #3115: Speeding up MNMG UMAP testing
 - PR #3112: Speed test_array
+- PR #3111: Adding Cython to Code Coverage
 
 ## Bug Fixes
 - PR #3065: Refactoring prims metrics function names from camelcase to underscore format